requirements.txt
asgiref==3.3.4
attrs==21.2.0
autopep8==1.5.7
cffi==1.14.5
cryptography==3.4.7
Django==3.2
django-filter==2.4.0
django-mysql==3.11.1
djangorestframework==3.12.4
jsonschema==3.2.0
Markdown==3.3.4
Pillow==8.2.0
pycodestyle==2.7.0
pycparser==2.20
PyMySQL==1.0.2
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
